認識されていなかったテストを追加
