+++ src/virtual_machine/parser.rs
@@ -1,7 +1,7 @@
 mod core;
-mod declaration_parser;
-mod expression_parser;
-mod parser_error;
+pub mod declaration_parser;
+pub mod expression_parser;
+pub mod parser_error;
 mod statement_parser;
 
 use crate::virtual_machine::ast::{Statement, AST};
