+++ src/virtual_machine/token.rs
@@ -3,7 +3,7 @@ pub mod token_type;
 use crate::virtual_machine::token::token_type::TokenType;
 
 #[derive(Clone, PartialEq, Debug)]
-pub(crate) struct Token {
+pub struct Token {
     pub(crate) line: usize,
     pub(crate) char_pos: usize,
     pub(crate) token_type: TokenType,
